scripts/requirements.txt
# Feature 1: Infrastructure Setup Scripts Dependencies
nats-py==2.10.0
qdrant-client==1.7.0
